requirements.txt
aiogram==3.0.0  # Для работы с Telegram Bot API
aiohttp==3.9.1  # Для асинхронных HTTP-запросов
aiosqlite==0.18.0  # Для асинхронной работы с SQLite
python-dotenv==1.0.1  # Для загрузки переменных окружения из файла .env
requests==2.31.0  # Для отправки HTTP-запросов (если используется)